chore: move StatefulConn to internal
